Code - Added getMenuById endpoint to ServiceController.

+++ src/main/java/backend/repository/IMenuRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IMenuRepository extends CrudRepository<Menu, Long> {
+public interface IMenuRepository extends CrudRepository<Menu, Integer> {
 }
